requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.0
djangorestframework==3.12.4
idna==3.3
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
